pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adamo"
version = "0.2.0"
description = "Python SDK for Adamo — publish and subscribe to data through Adamo's global Zenoh infrastructure"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "eclipse-zenoh>=1.7.2",
    "httpx>=0.28",
]

[project.optional-dependencies]
data = ["pandas>=2"]
ml = ["torch>=2", "av>=13", "numpy>=1.24"]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
]
